Move components folder into src/views/
