+++ src/views/components/CoursePeriodsTable.vue
