+++ src/views/components/TheFooter.vue
