+++ src/views/components/assignments/upcoming/AssignmentPanelBlock.vue
