+++ src/views/components/courses/CourseModalLinks.vue
