+++ src/views/components/modal/ModalSelectCourse.vue
