+++ src/views/components/sidebar/SidebarSchedule.vue
